bloc2-architecture-donnees/stripe-data-architecture/generator/requirements.txt
psycopg2-binary==2.9.9
faker==22.5.1
